+++ src/pages/MyInventories/MyInventories.js
@@ -3,7 +3,6 @@ import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import React, { useEffect, useState } from "react";
 import { Table } from "react-bootstrap";
 import { useAuthState } from "react-firebase-hooks/auth";
-import toast from "react-hot-toast";
 import { useNavigate } from "react-router-dom";
 import { auth } from "../../firebase.init";
 
@@ -21,35 +20,25 @@ const MyInventories = () => {
   }, []);
 
   const handleDeleteInventory = (id) => {
-    console.log(id);
-    fetch(`http://localhost:5000/deleteInventory?id=${id}`, {
-      method: "DELETE",
-    })
-      .then((res) => res.json())
-      .then((data) => {
-        if (data.deletedCount > 0) {
-          const restInventories = inventories.filter(
-            (inventory) => inventory._id !== id
-          );
-          setInventories(restInventories);
-        }
-      });
+    const confirm = window.confirm("You wanna delete?");
+    if (confirm) {
+      console.log(id);
+      fetch(`http://localhost:5000/deleteInventory?id=${id}`, {
+        method: "DELETE",
+      })
+        .then((res) => res.json())
+        .then((data) => {
+          if (data.deletedCount > 0) {
+            const restInventories = inventories.filter(
+              (inventory) => inventory._id !== id
+            );
+            setInventories(restInventories);
+          }
+        });
+    }
   };
 
   return (
-    // <div>
-    //   <h2 className="text-center">Your Inventories {inventories.length}</h2>
-    //   <div className="myInventoriesContainer">
-    //     {/* {inventories.map((inventory) => (
-    //       <div key={inventory._id}>
-    //         <h3>{inventory.name}</h3>
-    //         <button onClick={() => handleDeleteInventory(inventory._id)}>
-    //           X
-    //         </button>
-    //       </div>
-    //     ))} */}
-    //   </div>
-    // </div>
     <div className="manageInventoriesContainer container">
       <h2 className="text-center">YOUR INVENTORIES</h2>
